refactor: move to separate function
