+++ apps/api/app/Services/ServerStatus.ts
@@ -1,12 +1,15 @@
+import Config from '@ioc:Adonis/Core/Config'
 import Character from 'App/Models/Character'
 import Server from 'App/Models/Server'
 import isPortReachable from 'is-port-reachable'
+import { Socket } from 'socket.io'
+import { DefaultEventsMap } from 'socket.io/dist/typed-events'
 
-export async function getServerStatus(port: number, host: string) {
+export async function getServerPortStatus(port: number, host: string) {
   return (await isPortReachable(port, { host })) ? 'online' : 'offline'
 }
 
-export async function initServerStatus() {
+export async function getServerStatus() {
   const serverSettings = await Server.query()
     .where('id', 1)
     .preload('settings', (settingsQuery) => {
@@ -30,7 +33,7 @@ export async function initServerStatus() {
           ...serverStatus,
           ...{
             [server.name]: {
-              status: await getServerStatus(serverValue?.port, serverValue?.host),
+              status: await getServerPortStatus(serverValue?.port, serverValue?.host),
             },
           },
         }
@@ -44,4 +47,26 @@ export async function initServerStatus() {
       }
     },
   }
+}
+
+export async function ServerStatusSocket(
+  socket: Socket<DefaultEventsMap, DefaultEventsMap, DefaultEventsMap, any>
+) {
+  const server = await getServerStatus()
+  const refreshTime = Config.get('ragnarok.server.statusRefreshTimeMs', 1000)
+  let emitServerStatusTimer: string | number | NodeJS.Timeout | undefined
+
+  async function emitServerStatus() {
+    console.log('emitServerStatus')
+    socket.emit('server-status', await server.status())
+    emitServerStatusTimer = setTimeout(emitServerStatus, refreshTime)
+  }
+
+  // clear timer
+  socket.on('disconnect', function () {
+    console.log('disconnect')
+    clearTimeout(emitServerStatusTimer)
+  })
+
+  emitServerStatus()
 }
